cmath to math.h

+++ float.cpp
@@ -9,7 +9,7 @@
 #include <mutex>
 #include <condition_variable>
 #include <signal.h>
-#include <cmath>
+#include <math.h>
 
 std::mutex m ;
 std::condition_variable cv ;
@@ -58,9 +58,9 @@ void square_am_signal(float time, float frequency) {
     while (high_resolution_clock::now() < end) {
         mid = start + period / 2 ;
         reset = start + period ;
-        int i = 0;
+        double i = 0.0;
         while( high_resolution_clock::now() < mid ) {
-            *ptr = std::log(ptr + i)*2.0;
+            *ptr = log(ptr + i)*2.0;
             i++;
             if(i==limit) i=0;
         }
